Adding gs1:defaultLink and defaultLink*
Also uses 'proposed' as value for vs:term_status. Note the defn in the term status ontology is "the status of a vocabulary term, expressed as a short symbolic string; known values include 'unstable','testing', 'stable' and 'archaic'". 'Proposed' is none of those but seems acceptable and is what I mean here.